Microservices/Building Microservice Architecture/Django/TestAPI1/requirements.txt
gunicorn
django
djangorestframework
django-rest-auth
django-allauth
django-cors-headers